separate ops and queries
