+++ test/table_test.go
@@ -1,56 +1,13 @@
-package dynago_test
+package test
 
 import (
 	"context"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/eyebrow-fish/dynago"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"reflect"
 	"testing"
 )
-
-func TestCreateTable(t *testing.T) {
-	process := setupLocalDynamo()
-	defer func() { panicOnError(process.Kill()) }()
-
-	table, err := dynago.CreateTable("testTable", testTable{})
-	if err != nil {
-		t.Fatal("error creating table:", err)
-	}
-	if table == nil {
-		t.Fatal("table was nil")
-	}
-	if table.Name != "testTable" {
-		t.Fatal("table was not called testTable")
-	}
-	if table.Schema != (testTable{}) {
-		t.Fatal("table was not", testTable{})
-	}
-}
-
-func TestCreateTable_duplicate(t *testing.T) {
-	process := setupLocalDynamo()
-	defer func() { panicOnError(process.Kill()) }()
-
-	_, _ = dynago.CreateTable("testTable", testTable{})
-	_, err := dynago.CreateTable("testTable", testTable{})
-	if err == nil {
-		t.Fatal("expected an error to occur")
-	}
-}
-
-func TestCreateTable_noHash(t *testing.T) {
-	process := setupLocalDynamo()
-	defer func() { panicOnError(process.Kill()) }()
-
-	_, err := dynago.CreateTable("testTable", struct{}{})
-	if err == nil {
-		t.Fatal("expected an error to occur")
-	}
-}
 
 func TestNewTable(t *testing.T) {
 	process := setupLocalDynamo()
@@ -114,34 +71,4 @@ var (
 type testTable struct {
 	Id   int
 	Name string
-}
-
-func setupLocalDynamo() *os.Process {
-	homeDir, err := os.UserHomeDir()
-	panicOnError(err)
-
-	libDir := filepath.Join(homeDir, "dev", "dynamo-local-lib")
-
-	command := exec.Command(
-		"java",
-		"-Djava.library.path="+filepath.Join(libDir, "DynamoDBLocal_lib"),
-		"-jar",
-		filepath.Join(libDir, "DynamoDBLocal.jar"),
-		"-inMemory",
-	)
-
-	panicOnError(command.Start())
-	panicOnError(exec.Command("aws", "dynamodb", "list-tables", "--endpoint-url", "http://localhost:8000").Run())
-
-	dynago.UpdateOptions(testOptions)
-
-	return command.Process
-}
-
-func panicOnError(err error) {
-	if err == nil {
-		return
-	}
-
-	panic(err)
 }
